Move dynamic value to its own file
